Add unit tests for admission webhook rejecting semantic equals
The admission webhook uses the Compare function in the structured-merge-diff
project to compute the diff to prevent drift:
https://github.com/GoogleContainerTools/kpt-config-sync/blob/v1.16.0/pkg/webhook/fields.go#L54.

There is a known issue in the Compare function that the equality is not
sufficient:
https://github.com/kubernetes-sigs/structured-merge-diff/blob/v4.2.3/typed/typed.go#L134.

This causes the admission webhook to reject a request even if they're
semantically equal.

This commit adds unit tests for this scenario. When it is fixed, the
unit tests should be updated to reflect the change.